refactor(iso): extract runLoader helper for RPC-vs-direct-fn path selection
Both LoaderHost call sites (runReload + first-render) had near-verbatim
fetch-path-selector ladders, with the direct-fn branch additionally
holding the streaming-SSR registry plumbing. Move both to a single
runLoader(loaderRef, location, id, signal, callbacks) in loader-runner.ts.
Behavior unchanged.
